Update create-player-hazards.md
The for loop is a fail safe in the event of 'Players.PlayedAdded' is not functioning perfectly. If the player collects the coins before the for loop is executed, the count would be reset to zero.
To handle this issue the suggested code suggests if the player already has coins collected and if the count is nil then it is reset.